Added simple push notification;

+++ PushCampaignSystem/WebApi/Models/Visit.cs
@@ -2,10 +2,10 @@ namespace WebApi.Models
 {
     public class Visit
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public string DeviceId { get; set; }
 
-        public long PlaceId { get; set; }
+        public int PlaceId { get; set; }
     }
 }
